+++ apis/v1/news/index.ts
@@ -1,4 +1,4 @@
-import { getRequest, patchRequest, postRequest } from '@/apis';
+import { getRequest, postRequest } from '@/apis';
 import { FETCH_TAG_NEWS } from '@/constants/network';
 import { NewsPreviewList } from '@/types/news';
 import { PostSearchQueryParams } from '@/types/post';
@@ -11,8 +11,4 @@ export const getNewsPosts = (params: PostSearchQueryParams) =>
 
 export const postNews = async (formData: FormData) => {
   return postRequest('/v1/news', { body: formData, jsessionID: true }) as Promise<{ id: number }>;
-};
-
-export const patchNews = async (id: number, formData: FormData) => {
-  await patchRequest(`/v1/news/${id}`, { body: formData, jsessionID: true });
 };
