+++ apis/v1/notice/[id].ts
@@ -1,0 +1,15 @@
+import { deleteRequest, getRequest, patchRequest } from '@/apis';
+import { FETCH_TAG_NOTICE } from '@/constants/network';
+import { Notice } from '@/types/notice';
+import { PostSearchQueryParams } from '@/types/post';
+
+export const getNoticePostDetail = (id: number, params: PostSearchQueryParams) =>
+  getRequest(`/v1/notice/${id}`, params, {
+    next: { tags: [FETCH_TAG_NOTICE] },
+    jsessionID: true,
+  }) as Promise<Notice>;
+
+export const patchNotice = (id: number, formData: FormData) =>
+  patchRequest(`/v1/notice/${id}`, { body: formData, jsessionID: true });
+
+export const deleteNotice = (id: number) => deleteRequest(`/v1/notice/${id}`, { jsessionID: true });
